Have fen with static methods and file rename
