+++ src/main/java/View/BoardView.java
@@ -2,11 +2,11 @@ package View;
 
 import View.Contract.Listener;
 
-public class BoardSquaresView {
+public class BoardView {
 
     SquareView[][] squareView = new SquareView[8][8];
 
-    public BoardSquaresView(Listener listener) {
+    public BoardView(Listener listener) {
         for (int rank = 0; rank < 8; rank++) {
             for (int file = 0; file < 8; file++) {
                 squareView[file][rank] = new SquareView(file, rank, listener);
